change namespace SampleWebSite. fix Microsoft.Web.Infrastructure ref. change folers ApplicationDbInitializer.cs / IdentityModel.cs
