+++ .gitignore
@@ -293,3 +293,4 @@ __pycache__/
 *.btm.cs
 *.odx.cs
 *.xsd.cs
+/Src/SampleWebSite/MyConnectionStrings.config
