+++ Src/SampleWebSite/App_Start/BundleConfig.cs
@@ -1,6 +1,6 @@
 ﻿using System.Web.Optimization;
 
-namespace IdentitySample
+namespace SampleWebSite
 {
     public class BundleConfig
     {
