+++ Src/SampleWebSite/App_Start/Startup.Auth.cs
@@ -2,11 +2,11 @@
 using Microsoft.AspNet.Identity.Owin;
 using Microsoft.Owin;
 using Microsoft.Owin.Security.Cookies;
-using IdentitySample.Models;
 using Owin;
 using System;
+using SampleWebSite.Models;
 
-namespace IdentitySample {
+namespace SampleWebSite {
     public partial class Startup {
         // For more information on configuring authentication, please visit http://go.microsoft.com/fwlink/?LinkId=301864
         public void ConfigureAuth(IAppBuilder app) {
