+++ Src/SampleWebSite/Controllers/AccountController.cs
@@ -1,5 +1,4 @@
 ﻿using System.Globalization;
-using IdentitySample.Models;
 using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.Owin;
 using Microsoft.Owin.Security;
@@ -9,8 +8,10 @@ using System.Security.Claims;
 using System.Threading.Tasks;
 using System.Web;
 using System.Web.Mvc;
+using SampleWebSite.Models;
+using SampleWebSite.ViewModels;
 
-namespace IdentitySample.Controllers
+namespace SampleWebSite.Controllers
 {
     [Authorize]
     public class AccountController : Controller
