+++ Src/SampleWebSite/Controllers/ManageController.cs
@@ -1,13 +1,14 @@
-﻿using IdentitySample.Models;
-using Microsoft.AspNet.Identity;
+﻿using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.Owin;
 using Microsoft.Owin.Security;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Web;
 using System.Web.Mvc;
+using SampleWebSite.Models;
+using SampleWebSite.ViewModels;
 
-namespace IdentitySample.Controllers
+namespace SampleWebSite.Controllers
 {
     [Authorize]
     public class ManageController : Controller
