+++ Src/SampleWebSite/Controllers/UserAdminController.cs
@@ -1,5 +1,4 @@
-﻿using IdentitySample.Models;
-using Microsoft.AspNet.Identity;
+﻿using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.Owin;
 using System;
 using System.Collections.Generic;
@@ -9,8 +8,10 @@ using System.Net;
 using System.Threading.Tasks;
 using System.Web;
 using System.Web.Mvc;
+using SampleWebSite.Models;
+using SampleWebSite.ViewModels;
 
-namespace IdentitySample.Controllers
+namespace SampleWebSite.Controllers
 {
     [Authorize(Roles = "Admin")]
     public class UsersAdminController : Controller
