+++ Src/SampleWebSite/Global.asax.cs
@@ -1,10 +1,10 @@
-﻿using IdentitySample.Models;
+﻿using SampleWebSite.Models;
 using System.Data.Entity;
 using System.Web.Mvc;
 using System.Web.Optimization;
 using System.Web.Routing;
 
-namespace IdentitySample
+namespace SampleWebSite
 {
     // Note: For instructions on enabling IIS7 classic mode, 
     // visit http://go.microsoft.com/?LinkId=301868
