+++ Src/SampleWebSite/ViewModels/ManageViewModels.cs
@@ -1,9 +1,9 @@
-using Microsoft.AspNet.Identity;
-using Microsoft.Owin.Security;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using Microsoft.AspNet.Identity;
+using Microsoft.Owin.Security;
 
-namespace IdentitySample.Models {
+namespace SampleWebSite.ViewModels {
     public class IndexViewModel {
         public bool HasPassword { get; set; }
         public IList<UserLoginInfo> Logins { get; set; }
