+++ Src/SampleWebSite/Views/Account/ExternalLoginConfirmation.cshtml
@@ -1,4 +1,4 @@
-@model IdentitySample.Models.ExternalLoginConfirmationViewModel
+@model SampleWebSite.ViewModels.ExternalLoginConfirmationViewModel
 @{
     ViewBag.Title = "Register";
 }
