+++ Src/SampleWebSite/Views/Account/ForgotPassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model IdentitySample.Models.ForgotPasswordViewModel
+﻿@model SampleWebSite.ViewModels.ForgotPasswordViewModel
 @{
     ViewBag.Title = "Forgot your password?";
 }
