+++ Src/SampleWebSite/Views/Account/Login.cshtml
@@ -1,4 +1,5 @@
-@model IdentitySample.Models.LoginViewModel
+@using SampleWebSite.ViewModels
+@model SampleWebSite.ViewModels.LoginViewModel
 
 @{
     ViewBag.Title = "Log in";
@@ -51,7 +52,7 @@
     </div>
     <div class="col-md-4">
         <section id="socialLoginForm">
-            @Html.Partial("_ExternalLoginsListPartial", new IdentitySample.Models.ExternalLoginListViewModel { ReturnUrl = ViewBag.ReturnUrl })
+            @Html.Partial("_ExternalLoginsListPartial", new ExternalLoginListViewModel { ReturnUrl = ViewBag.ReturnUrl })
         </section>
     </div>
 </div>
