+++ Src/SampleWebSite/Views/Account/Register.cshtml
@@ -1,4 +1,4 @@
-@model IdentitySample.Models.RegisterViewModel
+@model SampleWebSite.ViewModels.RegisterViewModel
 @{
     ViewBag.Title = "Register";
 }
