+++ Src/SampleWebSite/Views/Account/ResetPassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model IdentitySample.Models.ResetPasswordViewModel
+﻿@model SampleWebSite.ViewModels.ResetPasswordViewModel
 @{
     ViewBag.Title = "Reset password";
 }
