+++ Src/SampleWebSite/Views/Account/SendCode.cshtml
@@ -1,4 +1,4 @@
-@model IdentitySample.Models.SendCodeViewModel
+@model SampleWebSite.ViewModels.SendCodeViewModel
 @{
     ViewBag.Title = "Send Verification Code";
 }
