+++ Src/SampleWebSite/Views/Account/VerifyCode.cshtml
@@ -1,4 +1,4 @@
-@model IdentitySample.Models.VerifyCodeViewModel
+@model SampleWebSite.ViewModels.VerifyCodeViewModel
 @{
     ViewBag.Title = "Enter Verification Code";
 }
