+++ Src/SampleWebSite/Views/Manage/AddPhoneNumber.cshtml
@@ -1,4 +1,4 @@
-@model IdentitySample.Models.AddPhoneNumberViewModel
+@model SampleWebSite.ViewModels.AddPhoneNumberViewModel
 @{
     ViewBag.Title = "Add Phone Number";
 }
