+++ Src/SampleWebSite/Views/Manage/ChangePassword.cshtml
@@ -1,4 +1,4 @@
-@model IdentitySample.Models.ChangePasswordViewModel
+@model SampleWebSite.ViewModels.ChangePasswordViewModel
 @{
     ViewBag.Title = "Change Password";
 }
