+++ Src/SampleWebSite/Views/Manage/Index.cshtml
@@ -1,4 +1,4 @@
-@model IdentitySample.Models.IndexViewModel
+@model SampleWebSite.ViewModels.IndexViewModel
 @{
     ViewBag.Title = "Manage your account";
 }
