+++ Src/SampleWebSite/Views/Manage/ManageLogins.cshtml
@@ -1,4 +1,4 @@
-@model IdentitySample.Models.ManageLoginsViewModel
+@model SampleWebSite.ViewModels.ManageLoginsViewModel
 @using Microsoft.Owin.Security
 @{
     ViewBag.Title = "Manage your external logins";
