+++ Src/SampleWebSite/Views/Manage/SetPassword.cshtml
@@ -1,4 +1,4 @@
-@model IdentitySample.Models.SetPasswordViewModel
+@model SampleWebSite.ViewModels.SetPasswordViewModel
 
 <p class="text-info">
     You do not have a local username/password for this site. Add a local
