+++ Src/SampleWebSite/Views/Manage/VerifyPhoneNumber.cshtml
@@ -1,4 +1,4 @@
-@model IdentitySample.Models.VerifyPhoneNumberViewModel
+@model SampleWebSite.ViewModels.VerifyPhoneNumberViewModel
 @{
     ViewBag.Title = "Verify Phone Number";
 }
