+++ Src/SampleWebSite/Views/RolesAdmin/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model IdentitySample.Models.RoleViewModel
+﻿@model SampleWebSite.ViewModels.RoleViewModel
 
 @{
     ViewBag.Title = "Create";
