+++ Src/SampleWebSite/Views/UsersAdmin/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model IdentitySample.Models.RegisterViewModel
+﻿@model SampleWebSite.ViewModels.RegisterViewModel
 @{
     ViewBag.Title = "Create";
 }
