+++ Src/SampleWebSite/Views/UsersAdmin/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model IdentitySample.Models.ApplicationUser
+﻿@model SampleWebSite.Models.ApplicationUser
 
 @{
     ViewBag.Title = "Details";
