+++ Src/SampleWebSite/Views/UsersAdmin/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model IdentitySample.Models.EditUserViewModel
+﻿@model SampleWebSite.ViewModels.EditUserViewModel
 
 @{
     ViewBag.Title = "Edit";
