+++ Src/SampleWebSite/Views/UsersAdmin/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<IdentitySample.Models.ApplicationUser>
+﻿@model IEnumerable<SampleWebSite.Models.ApplicationUser>
 
 @{
     ViewBag.Title = "Index";
