+++ Src/SampleWebSite/Web.config
@@ -4,9 +4,9 @@
   http://go.microsoft.com/fwlink/?LinkId=169433 を参照してください
   -->
 <configuration>
-  <connectionStrings configSource="MyConnectionStrings.config"/>
+  <connectionStrings configSource="MyConnectionStrings.config" />
   <appSettings>
-    <add key="owin:AppStartup" value="IdentitySample.Startup,SampleWebSite" />
+    <add key="owin:AppStartup" value="SampleWebSite.Startup,SampleWebSite" />
     <add key="webpages:Version" value="3.0.0.0" />
     <add key="webpages:Enabled" value="false" />
     <add key="ClientValidationEnabled" value="true" />
